adding is book borrowed method in book service

+++ src/main/java/com/shubh/library_management/service/BookServiceimpl.java
@@ -2,9 +2,11 @@ package com.shubh.library_management.service;
 
 import java.util.List;
 
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
 import com.shubh.library_management.dto.BookDTO;
+import com.shubh.library_management.dto.BorrowedBookDTO;
 import com.shubh.library_management.entity.Book;
 import com.shubh.library_management.mapper.BookMapper;
 import com.shubh.library_management.repository.BookRepository;
@@ -13,8 +15,11 @@ import com.shubh.library_management.repository.BookRepository;
 public class BookServiceimpl implements BookService{
     private BookRepository bookRepository;
 
-    public BookServiceimpl(BookRepository bookRepository) {
+    private BorrowedBookService borrowedBookService;
+
+    public BookServiceimpl(BookRepository bookRepository, @Lazy BorrowedBookService borrowedBookService) {
         this.bookRepository = bookRepository;
+        this.borrowedBookService = borrowedBookService;
     }
 
     public List<BookDTO> getAllBooks(){
@@ -39,7 +44,27 @@ public class BookServiceimpl implements BookService{
         return BookMapper.mapToBookDTO(bookRepository.save(book));
     }
 
-    public void deleteBook(Long bookId){
-        bookRepository.deleteById(bookId);
+    public String deleteBook(Long bookId){
+        if (isBookBorrowed(bookId)) {
+            bookRepository.deleteById(bookId);
+            return "book deleted successfully";
+        }else{
+            return "book will be deleted when all the user's have returned the book.";
+        }
+    }
+
+    public boolean isBookBorrowed(Long bookId){
+        List<BorrowedBookDTO> borrowedBooks = borrowedBookService.getAllBorrowedBooks();
+
+        for (BorrowedBookDTO borrowedBookDTO : borrowedBooks) {
+            if (borrowedBookDTO.getBookId() == bookId && borrowedBookDTO.isReturned() == false) {
+                Book book = getBookbyId(bookId);
+                book.setBookCount(0);
+                book.setSetToDeleted(true);
+                updateBook(book);
+                return false;
+            }
+        }
+        return true;
     }
 }
